Fixed bad scaling bug.
For long perfpoint runs, the ipoint value can get *very* large.
The scaled alignment uses a 64-bit floating point number to store the scaling
factor. With a large ipoint value, the difference between the last desired
ipoint and the computed scaled value approaches the precision of a double.

For example, a last ipoint of 851880000000 is approximately 2^39.
The machine epsilon is 2^-52, which gives a ULP of 2^-13, or 0.000122.
This is larger than the threshold we were using (0.0001).
I bumped it by five orders of magnitude.
This should allow 4,503,599,627,370,496 instructions, or about 17 days of
runtime at 3 giga-insns per second.